Updates to data tables
Added an edited data table I found with Suemitsu's name in it, changed it. It did not change the name in the social link menu though, so I'm unsure of what it actually changed. Oh well, I'm sure it's good that got added anyways.

Edited the previous data file with Suemitsu's name to change some space characters (hex 20) to null characters (hex 00).